rename google_map.vue to GoogleMap.vue
